refactor: Header, Sidebar 경로 변경
