+++ src/widgets/ui/Header.tsx
@@ -2,7 +2,7 @@ import { Bell, KeyRound, Lock, CirclePlus } from 'lucide-react';
 import { useMemo } from 'react';
 import { Link } from 'react-router';
 
-import { ROUTES } from '../app/routes/routes';
+import { ROUTES } from '../../app/routes/routes';
 import logo from '../shared/assets/logo/Megabox_Logo_Indigo.png';
 
 export const Header = () => {
